--- a/make-on-hpc.sh
+++ b/make-on-hpc.sh
@@ -12,7 +12,7 @@
 # Loading the required modules
 module load Java
 module load Maven
-module load node
+module load nodejs
 
 # Check the current directory
 pushd "$PBS_O_WORKDIR"